power: Add Google USB-C to 3.5mm device to autosuspend whitelist
Google's USB-C to 3.5mm Digital Headphone Jack Adapter ("Mir") is available:
https://store.google.com/product/usb_c_headphone_adapter

The dongle may be attached to Chromebooks.
We want to exercise the low power mode of this device, so let's
whitelist it for USB autosuspend.

Signed-off-by: Benson Leung <bleung@chromium.org>

BUG=b:64950327
TEST=Attach Mir USB-C to 3.5mm device to Electro/Reef. Check in
/sys/bus/usb/devices that the device in question's power/runtime_status
settles on "Suspended" and power consumption is ~2.5mA as measured by
an inline meter.

Change-Id: Ibbebb7396b7e2b7b4f4c8b2403b02ac0ff7b3edc
Reviewed-on: https://chromium-review.googlesource.com/732729
Commit-Ready: Benson Leung <bleung@chromium.org>
Tested-by: Benson Leung <bleung@chromium.org>
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>