init: Erase more key material in clobber-state
Add "encrypted.needs-finalization" to the list of keysets to wipe for
the eMMC fast wipe code path. Still the problem remains that this file
cannot be securely erased anymore after it has been deleted from the
filesystem as part of "finalization" in mount-encrypted.c.

BUG=chromium:758563
TEST=manual

Change-Id: I4ac2cb2cfee259007e97976a5fb0003cbacf816d
Reviewed-on: https://chromium-review.googlesource.com/632619
Commit-Ready: Thiemo Nagel <tnagel@chromium.org>
Tested-by: Thiemo Nagel <tnagel@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>