shill: Change IPConfig::Route to use prefixes instead of netmasks
Most users call IPAddress::GetPrefixLengthFromMask() to convert the
string to an int.  And that function only works for IPv4 anyway.
Store it as an int "natively" to simplify the code.

Minor downside: if openvpn pushes a route that is missing the netmask,
that condition will not be detected.  That shouldn't happen anyway,
because route_{network,netmask,gateway} are all set from the same
function in the openvpn code.

BUG=chromium:696865
TEST=unit tests

Change-Id: I7049bc81481c7bebb267d0a242cc7e3049a7a93d
Reviewed-on: https://chromium-review.googlesource.com/636156
Commit-Ready: Kevin Cernekee <cernekee@chromium.org>
Tested-by: Kevin Cernekee <cernekee@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>