login: Flag for devices transitioning from ARC M to ARC N
There's a group of devices that had ARC M in version 60. The
devices from that group that are in EDU require a separate logic
when the migration policy is not set by admin. This flag allows
the device to recognize the group it is in. Design doc for this:
https://docs.google.com/document/d/1Wi5K1rTu74kHxznbaKGOP8fALfJsgPhvb5uA9K9AVPg/edit?ts=59a1f193#

BUG=chromium:722371
TEST=None

Change-Id: Ie0b95a48146cfebe3b25d72f4a3ef770feba1fc6
Reviewed-on: https://chromium-review.googlesource.com/637771
Commit-Ready: Igor <igorcov@chromium.org>
Tested-by: Igor <igorcov@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>