init: add a helper for waiting for unix sockets
Currently our syslog init script returns before it's fully available
(namely, apps can syslog() and not lose messages).  We need a method
for polling the /dev/log UNIX socket and not returning until it's
usable.  Hence, this simple tool.

It's written in such a way that we could conceivably extend it to
TCP/UDP ports if the need arises.

BUG=chromium:702794
TEST=precq passes

Change-Id: I6be60cee4c63289fc2a091b1ff6a3a44ff65364c
Reviewed-on: https://chromium-review.googlesource.com/639090
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Mattias Nissler <mnissler@chromium.org>