chromeos-config: Support shared configuration for white-label
White-label devices are almost electronically identical and need to share
most configuration, except a few things like the key ID and wallpaper.
Repeating the configuration is error-prone and wastes time.

Provide a way to share the configuration from another model. Local
settings and then not permitted, avoiding potential mistakes.

The approach here is that a single whitelabel model is created with the
shared configuration for all whitelabel models, and the other whitelabel
models share it. It is possible to have two sets of whitelabels if needed.

Validation is provided to ensure that whitelabel models do not specify
properties which they are supposed to share with other whitelabels. Any
property or node lookups return the correct information, taking into
account the whitelabel share.

This is all in one big CL so that the various implementations are
together for review. We should consider spltting it.

BUG=chromium:761229
TEST=FEATURES=test emerge-coral chromeos-config-tools
PYTHONPATH=~/cosarm ./validate/validate_config \
~/cosarm/chroot/build/coral/usr/share/chromeos-config/config.dtb  \
~/cosarm/chroot/build/reef-uni/usr/share/chromeos-config/config.dtb \
README.md
PYTHONPATH=~/cosarm python validate/validate_config_unittest.py

Change-Id: Ia9a2c5a1d203b29d96442cfd6748697521e6b35c
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/646995
Reviewed-by: Jason Clinton <jclinton@chromium.org>