shill: cellular: remove ControlInterface argument in ModemManager constructor
The constructor of ModemManager (and its subclasses) take both a
ControlInterface object and a ModemInfo object as arguments. As the same
ControlInterface object is accessible from the ModemInfo object, there
is no need to pass the ControlInterface object as an constructor
argument to ModemManager.

BUG=None
TEST=Run unit tests.

Change-Id: I254dbe197958b29829b4dba6c2d86febed9aebb4
Reviewed-on: https://chromium-review.googlesource.com/648647
Commit-Ready: Ben Chan <benchan@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>