shill: cellular: remove ControlInterface argument in Modem constructor
The constructor of Modem (and its subclasses) take both a
ControlInterface object and a ModemInfo object as arguments. As the same
ControlInterface object is accessible from the ModemInfo object, there
is no need to pass the ControlInterface object as an constructor
argument to Modem.

BUG=None
TEST=Run unit tests.

Change-Id: I22593eeb37ee13ba23c1734336614b3780357e5d
Reviewed-on: https://chromium-review.googlesource.com/648648
Commit-Ready: Ben Chan <benchan@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>