shill: cellular: remove ControlInterface argument in CellularCapability constructor
The constructor of CellularCapability (and its subclasses) take both a
ControlInterface object and a ModemInfo object as arguments. As the same
ControlInterface object is accessible from the ModemInfo object, there
is no need to pass the ControlInterface object as an constructor
argument to CellularCapability.

BUG=None
TEST=Run unit tests.

Change-Id: Ibe050d85523856592c926f248964d88915d66065
Reviewed-on: https://chromium-review.googlesource.com/648649
Commit-Ready: Ben Chan <benchan@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>