hammerd: Force update at boot and show boot message during update.
On boot, we try to update several kinds of firmware. If the firmware
needs to update, we block the UI at until update finishes, and then go
back to Chrome UI. Therefore, we need a new upstart job to check these
jobs sequentially.

This CL adds a job "boot-update-firmware" for this. It triggers
hammerd and show up the message if the hammer firmware needs update.

When the base does not need to update (firmware is up to date, or the
base is not attached), the task only takes 0.06 second. It does not
affect the bootup time significantly.

BUG=b:65180379, chromium:764232
TEST=With new firmware, plug hammer and reboot poppy,
     check the boot message shows up and
     keyboard works after entering UI
TEST=With original firmware, plug hammer and reboot poppy,
     check the boot message does not show up, enter UI directly
TEST=With new firmware, unplug hammer and reboot poppy,
     check the boot message does not how up, enter UI directly.
     Then plug hammer, check the keyboard works after 5 seconds.
CQ-DEPEND=CL:651711,CL:656859

Change-Id: If5324c036faa8a7520f06945871a0d1b602d4da3
Reviewed-on: https://chromium-review.googlesource.com/651788
Commit-Ready: Chih-Yu Huang <akahuang@chromium.org>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>