cros-disks: Do not overwrite capabilities of the rename process
Because cros-disks runs as a non-privileged process with limited
permissions, we are not able to change capabilities of a child
process. Due to limited permissions, minijail will fail when it
tries to prctl(PR_SET_KEEPCAPS). As a result of the CL, we do not
overwrite capabilities of the rename process anymore.

BUG=chromium:761775
TEST=platform_CrosDisksRename

Change-Id: I1e2ae111b5fc437710ffd16c61c0a2aae722124d
Reviewed-on: https://chromium-review.googlesource.com/649588
Commit-Ready: Ben Chan <benchan@chromium.org>
Tested-by: Klemen Kozjek <klemenko@google.com>
Reviewed-by: Ben Chan <benchan@chromium.org>