midis: Add class to handle Mojo client calls
These are invoked from the ARC client.
When MidisHost.Connect() is called, a new client should be registered.
The MidiServerImpl class will have a handle to the correponding
Client object in the server, so it can forward all actions to the actual
client. The MidisServerImpl will also have a handle to the MidisClient
interface defined by the client, so that it can easily communicate
device information back to the client.

For the time being, we leave the existing method of creating clients,
via Unix Domain Sockets.

In the following patch, the functionality of the new mojo interfaces will
be fleshed out, and the Unix Domain support will be dropped.

BUG=chromium:701005
TEST=ARC client can register a client connection with midis, call
methods.

Change-Id: I8498bbb5e929b630fd534d93ecb0898eb8584035
Reviewed-on: https://chromium-review.googlesource.com/653363
Commit-Ready: Prashant Malani <pmalani@google.com>
Tested-by: Prashant Malani <pmalani@google.com>
Reviewed-by: Luis Hector Chavez <lhchavez@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>