run_oci/crosh: Support detaching when running
This change introduces a new way of running run_oci, in which it can
either keep running until init exits (with the 'run' command, the old
behavior) or detach as soon as the post-start hooks have run (with the
'start' command, similar to how runc works). This change also adds the
'kill' and 'destroy' commands, with similar semantics to runc's.

BUG=b:65462041
TEST=cros_workon_make --board=${BOARD} --test run_oci
CQ-DEPEND=CL:665960

Change-Id: I29d22d2ad5363d52ea43e2dabc4eaa1b601bc200
Reviewed-on: https://chromium-review.googlesource.com/656510
Commit-Ready: Luis Hector Chavez <lhchavez@chromium.org>
Tested-by: Luis Hector Chavez <lhchavez@chromium.org>
Reviewed-by: Luis Hector Chavez <lhchavez@chromium.org>