power: Add get_powerd_initial_backlight_level conversions.
Add --level_to_percent and --percent_to_level flags to the
get_powerd_initial_backlight_level tool to convert between
brightness levels and nonlinear percents. This can be
helpful when configuring the default brightness for new
devices.

This is a bit of an abuse of the existing tool, as the new
behavior can print a percent rather than a level and doesn't
have anything to do with the initial value, but it feels
better than duplicating the existing code for initializing a
BacklightController into yet another tool.

BUG=none
TEST=manual

Change-Id: If60dd64864e7b5365793514a4b490401b4981405
Reviewed-on: https://chromium-review.googlesource.com/658320
Commit-Ready: Dan Erat <derat@chromium.org>
Tested-by: Dan Erat <derat@chromium.org>
Reviewed-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>