vm_tools: maitred: Add support to wait for processes
When setting up the environment for running containers we need to
perform some steps serially and we need to ensure that each step
completed successfully before continuing on to the next one.

To address this, add a new option to the LaunchProcess RPC to wait until
the launched process exits.  If this option is set the RPC will also
return information about how the process exited (exited normally or
killed by signal).

BUG=none
TEST=Spawn some processes with the wait_for_exit flag set and see that
     everything works.

Change-Id: Ia0edd5c3c4ec96a55812f9a4cb438ee618399465
Signed-off-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/658143
Reviewed-by: Stephen Barber <smbarber@chromium.org>