shill: ignore IP config updates on stub devices
DeviceStub, a subclass of Device, is created for a non-operable or
blacklisted device that is not managed by shill. Upon invocation of
DeviceStub::OnIPConfigUpdated() or DeviceStub::OnIPv6ConfigUpdated() due
to an IP config update, a Connection object is created and owned by
DeviceStub. But as a DeviceStub object isn't registered with Manager,
Manager::DeregisterDevice() won't be called on the DeviceStub object.
The Connection object owned by the DeviceStub object won't be destroyed
in the expected order, which leads the crash described in b:62137021.

This CL makes DeviceStub::OnIPConfigUpdated() and
DeviceStub::OnIPv6ConfigUpdated() a no-op, such that no Connection
object is created upon an IP config update.

BUG=b:62137021
TEST=Run unit tests.

Change-Id: I2ef237f6d973ff22b5adf3cdeef4ac9bf120624e
Reviewed-on: https://chromium-review.googlesource.com/664919
Commit-Ready: Ben Chan <benchan@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>