vm_tools: maitred: Improve child process error handling
Right now if a failure occurs in one of the steps that run in the child
process after it forks but before it executes the requested program we
don't have any way of finding out about it.  The only thing the parent
process sees is that the child exited with a non-zero status.

Improve the error handling by using a socketpair to communicate any
setup failures back to the parent process.  After forking, the parent
process blocks on a read on one end of the socketpair while the child
process does its setup.  If any step fails, the child process writes a
struct with the error details to its end of the socketpair.  If all
steps succeed the child process will not write anything to the socket
and its end will be closed on a successful exec because the socketpair
is created with the SOCK_CLOEXEC flag.  This means that if the parent
reads 0 bytes from the socket then the child process successfully
executed the requested program.

BUG=none
TEST=manual

Change-Id: I8a4b1c4a5f5c8054464649e1609382a504d65a7f
Signed-off-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/668094
Reviewed-by: Dylan Reid <dgreid@chromium.org>