chromeos-config: Minimalist refactoring cros_config to run on the host
This is a refactor to libcros_config to allow it to run on the host via
a InitOnHost method that takes in a path to a config database and
optionally a model name. Also added a cros_config_host CLI that can also
be used to list models in a config database.

BUG=chromium:761264
TEST=Ran all unit tests including newly added

Change-Id: I3d7c563aa358283596127efd76ada7f6b95c5a0d
Reviewed-on: https://chromium-review.googlesource.com/666479
Commit-Ready: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>