midis: Mojo interface functions - part 1
This CL implements some of the functions for both the MidisServer
and MidisClient interfaces.

In the process of doing the above, we remove the corresponding
ListDevices() and DeviceAdded/DeviceRemoved() functionality for devices
which were registered through Unix Domain Sockets.

BUG=chromium:701005
TEST=Boot an NYC ARC instance on samus and verify that the
callbacks being added are called correctly.
Change-Id: Id11230622a689e9493a9fba210a663a8b06f3459
Reviewed-on: https://chromium-review.googlesource.com/668170
Commit-Ready: Prashant Malani <pmalani@google.com>
Tested-by: Prashant Malani <pmalani@google.com>
Reviewed-by: Ken Rockot <rockot@chromium.org>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>