installer: change postinst to run inside the installed path
All the postinst logic manually adds the specified path to the prefix of
the tools it runs in order to find the right binary.  But we run into the
problem of running dynamic programs against a different runtime.  To work
around this, we statically linked things, but that's problematic:
- we sometimes forget to statically link programs
- it's hard to know what all programs actually need to be
- shell scripts make this problem much worse
- mixing of static & dynamic causes disk bloat
- static linkage bleeds into other targets too (firmware updater/initramfs)

Lets change the postinst code so that it runs in the target root as any
other chroot.  All of our dynamic linkage issues go away as a result.

BUG=chromium:765753
TEST=precq passes
TEST=recovery image passes
TEST=cros flash passes

Change-Id: I9d3fa9b6f419ada75e8f94b9dd0260fcef486b64
Reviewed-on: https://chromium-review.googlesource.com/669013
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>