shill: cellular: observe hardware revision property updates from ModemManager
ModemManager has added a 'HardwareRevision' property to the Modem
interface for reporting the hardware revision of a MBIM or QMI modem.
This CL modifies CellularCapabilityUniversal to observe the
'HardwareRevision' property from a modem object exposed by ModemManager
and propagate that information to the Cellular.HardwareRevision property
of the device object exposed by shill.

BUG=chromium:766178
TEST=Verify that shill reports the hardware revision of a MBIM modem
     through the Cellular.HardwareRevision property of the device object
     for the modem.

Change-Id: Ida794a89c695f6e08923caa17dfc1a6dc7f3b517
Reviewed-on: https://chromium-review.googlesource.com/671304
Commit-Ready: Ben Chan <benchan@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>