init: Add sync to chromeos_shutdown
Sync before unmounting in chromeos_shutdown to guard against present and
(potential) future unmount failures.

(This is intentionally placed before writing boot stats because the sync
time is legitimately part of shutdown and should be covered by the
stats. Due to the change in coverage, this CL will look like a
regression though it likely isn't one as the subsequent umount will be
sped up by approximately the amount of time that sync takes.)

BUG=chromium:760007
TEST=manual: reboot works

Change-Id: Ic5e577f3facfd3f1b12862b211248c387fb65210
Reviewed-on: https://chromium-review.googlesource.com/671244
Commit-Ready: Thiemo Nagel <tnagel@chromium.org>
Tested-by: Thiemo Nagel <tnagel@chromium.org>
Reviewed-by: Gwendal Grignou <gwendal@google.com>