Add proto for policy storage and retrieval
Adds a proto that is used to simplify the policy storage and
retrieval D-Bus interface of Session Manager. The interface is
starting to explode combinatorically with
- 5 StorePolicy methods (7 with extension policy, see crbug.com/735100) and
- 4 RetrievePolicy methods (6 with extension policy).
The new interface is flexible and has only 3 methods
- StorePolicy
- StoreUnsignedPolicy and
- RetrievePolicy
that each take the proto from this CL to specify the kind of policy
to store or retrieve.

BUG=chromium:765644
TEST=emerge-amd64-generic system_api

Change-Id: I4affbec9b4ce921db21861d28e096d935aada2e6
Reviewed-on: https://chromium-review.googlesource.com/671268
Commit-Ready: Lutz Justen <ljusten@chromium.org>
Tested-by: Lutz Justen <ljusten@chromium.org>
Reviewed-by: Lutz Justen <ljusten@chromium.org>