login: Add simplified policy storage interface
Session Manager's D-Bus interface for policy storage and retrieval is
starting to explode combinatorically with
- 5 StorePolicy methods (7 with extension policy, see crbug.com/735100) and
- 4 RetrievePolicy methods (6 with extension policy).
This CL adds a new interface that is flexible and only has 3 methods
- StorePolicy
- StoreUnsignedPolicy and
- RetrievePolicy
that each take the proto from CL:671268 to specify the kind of policy
to store or retrieve.

Note that the old interface cannot be simply removed since it is
still used (at least) by Chrome and authpolicyd. It will be phased
out as soon as all users are swapped over to the new interface.

CQ-DEPEND=CL:671268

BUG=chromium:765644
TEST=emerge-amd64-generic chromeos-login
     cros_run_unit_tests --board=amd64-generic --packages chromeos-login

Change-Id: If4160b11ac363879282095e4c227dc940615162a
Reviewed-on: https://chromium-review.googlesource.com/671234
Commit-Ready: Lutz Justen <ljusten@chromium.org>
Tested-by: Lutz Justen <ljusten@chromium.org>
Reviewed-by: Mattias Nissler <mnissler@chromium.org>