crosh: allow programs to supply help/usage information
There's a better --help listing in some of the scripts we maintain, but
we force people to either duplicate the content in crosh, or run yet a
different command to get the better help ("help" instead of "--help").

Allow commands to register a help_$cmd function which will take care of
displaying all the extended details needed.

BUG=none
TEST=`help network_diag`, `network_diag -h`, `network_diag --help`

Change-Id: I32446b2023c6c647cc460f8a1d6382e4ff875531
Reviewed-on: https://chromium-review.googlesource.com/671877
Commit-Ready: Brian Norris <briannorris@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>