hammerd: Add field allowed_fw_hash and use it to validate
CL:641736 introduced a new field in update_fw.h on EC side. We need to modify
on our side so the communication will still hold.

After we have that information. We check if the SHA value of the entire
touchpad firmware has same hash against the record in RW firmware. This sanity
check prior to update because if an individual chunk verification fail, the
touchpad will stuck into a weird state (only part of the flash is updated).

BUG=b:63993173
TEST=FEATURES=test emerge-soraka hammerd
TEST=With staff_v1.1.7056-73b314c1d by
     make TOUCHPAD_FW=SA461D-10H2_Fv01.bin BOARD=staff -j
     successfully update the touchpad firmware.
TEST=Manually replace the firmware to SA461D-10H2_Fv02.bin.
     Log shows it stopped at the stage of entire firmware hash check.

Change-Id: I894f3022c960b62c39a3f212737edb6a5d213769
Reviewed-on: https://chromium-review.googlesource.com/672326
Commit-Ready: Chun-ta Lin <itspeter@chromium.org>
Tested-by: Chun-ta Lin <itspeter@chromium.org>
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>