libcontainer: Create device nodes in the container's mount namespace
When setting up a container, /dev will most likely be a tmpfs that's
mounted inside the container and / might come from a read-only
filesystem, so any mknod(2) calls outside that will fail due to the path
not being writable in the outer namespace. Furthermore, just running a
hook in the container's process might not be sufficient, because that
process might be running in a user namespace, which would make the calls
to mknod(2) fail due to not having the CAP_MKNOD capability in the init
namespace.

This change moves the device node creation to a stage after the
container has finished setting up the mount namespace but before it
calls chroot(2). This is done in the init's user namespace (so it has
the correct capabilities to call mknod(2)) and the container's mount
namespace (so it can access the right path).

CQ-DEPEND=CL:678118,CL:695385,CL:*469812
BUG=b:65462041
TEST=cros_workon_make --board=${BOARD} --test run_oci
TEST=bertha still boots

Change-Id: Idc2c069a6b4b4658e4fae1dc9239067d2393d47f
Reviewed-on: https://chromium-review.googlesource.com/675842
Commit-Ready: Luis Hector Chavez <lhchavez@chromium.org>
Tested-by: Luis Hector Chavez <lhchavez@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>