crosh: switch shill commands to run via debugd
The majority of the commands just make dbus requests and don't need
access to local resources.  Only stand out is the network_diag util
as it writes its log output to the Downloads folder.  Move that out
of the script and into crosh so we can preserve behavior.

BUG=chromium:766379
TEST=running from crosh works

Change-Id: I8405d4c8470e2a41b62c5b20ede620274aad3ba0
Reviewed-on: https://chromium-review.googlesource.com/675734
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>