hammerd: Update the pack_firmware script.
This CL implements:
- Download the touchpad firmware.
- Download hammer EC firmware with premp and mp key with version.
  e.g. premp-v2
  Note that we should explictly assign which version of key do we use.

Because we fetch the EC firmware and touchpad firmware from the same
version and pack to a single tarball, we might face the situation that
only touchpad firmware is upgraded. Therefore we cannot use EC
firmware version as the tarball file name. Now the file naming is:
  <Base name>_<version>_<signed_key>.tbz2
For example:
  hammer_9959.0.0_premp-v2.tbz2

BUG=b:65230801
TEST=Run `./pack_firmware.sh -b poppy -v 9959.0.0 -s premp-v2`
     and check the tarball is generated

Change-Id: I52bbf47d4b7230e0bf6fdf8742e12ba878d6c690
Reviewed-on: https://chromium-review.googlesource.com/674514
Commit-Ready: Chih-Yu Huang <akahuang@chromium.org>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>