power: use RTC-relative wakeup timeouts for suspend
Most of our suspend/resume testing tools (powerd_suspend and
suspend_stress_test) use the relative form to set wakealarms (write "+X"
to '/sys/class/rtc/rtc0/wakealarm', where 'X' is the time offset). This
allows the kernel to determine the RTC time to pass down to the device.

However, powerd_dbus_suspend for some reason computes an absolute time,
using the current system time. This mostly works fine if the system
clock and the RTC are closely in sync, but it creates problems if the
RTC is significantly leading/lagging.

Let's be consistent, and just use the relative form.

Noticed while bringing up an EC-based RTC on STM32 (CL:674054).

BUG=b:63908519
TEST=`powerd_dbus_suspend --wakeup_timeout=10`

Change-Id: I2d39f7160dcd85ef9c6c49fa80647289f41e4279
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/677625
Reviewed-by: Dan Erat <derat@chromium.org>