vm_tools: syslog: Don't repeat kernel logs
If vm_syslog restarts for any reason it will re-send all the VM kernel's
logs to the LogCollector service.  This can create a lot of noise in the
system logs.

Luckily the /dev/kmsg interface provides a sequence number for each log
message.  Store this number in a special file in the /run directory and
update it every time new kernel logs are flushed.  On startup, check if
this file exists and if it does, don't send any kernel logs whose
sequence number is less than or equal to the sequence number in the
file.

BUG=none
TEST=unit tests and manual: restart vm_syslog and verify that
     /var/log/messages on the host does not fill up with repeated kernel
     log messages from the VM

Change-Id: If5a12569bc4067da876a8f344f64148256cdb5e1
Signed-off-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/685439
Reviewed-by: Stephen Barber <smbarber@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>