vm_tools: syslog: Log to /dev/kmsg
Currently all the logs from vm_syslog are being dropped.  Sending logs
to /dev/log also doesn't work for early setup before the socket is
ready.  We could buffer those logs and send them later but that's more
complicated.  We could also set up a loopback and automatically buffer
self-logs.  The main disadvantage is that we would lose any logs related
to failures in sending messages to the LogCollector service, which is
arguably the most important time when logging should work.

Instead just send all the logs to /dev/kmsg.  The main justification is
this ensures that vm_syslog logs will appear in the kernel log buffer
and don't depend on grpc or anything other than the guest kernel.  Also,
the code already exists and can easily be copied from main.cc in
maitred.  If there's ever a 3rd application that logs to
/dev/kmsg (unlikely), we should refactor the code into a single
location.

BUG=none
TEST=start a VM and see that vm_syslog logs appear in the kernel log
     buffer

Change-Id: Id793546086b6bdf93b31c57a08d5dd9a3a0fae77
Signed-off-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/685440
Reviewed-by: Mike Frysinger <vapier@chromium.org>