crosh: Add a command to start a container in a VM
The new "vmc" command is used to start a container in a virtual machine.
This is added separately from the existing "c" container command that
starts google-signed containers directly. Running containers in a VM
allows for running untrusted containers.

VMs will only start on boards built with the kvm_host USE flag set.
Don't try to start VMs on systems that lack the vm tools.

BUG=703958
TEST=crosh> vmc start <container path relative to Downloads> test_vm

Change-Id: I12b0f25cef876320b285a2a2c3792ef41bd73438
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/688096
Reviewed-by: Mike Frysinger <vapier@chromium.org>