hammerd: split updatepossible in preparation for metrics
It doesn't actually make much sense for UpdatePossible to
take an argument (RO or RW) since it is only ever used for RW.
Remove the argument, and split into ValidKey and ValidRollback.
This gives more fine-grained information which will be
used for UMA metrics.

BUG=b:64195183
TEST=cros_workon_make --board=poppy hammerd --test
TEST=USE=-cros-debug emerge-poppy -v --nodeps hammerd chromeos-bsp-poppy \
     && cros deploy --force --board=poppy dut hammerd chromeos-bsp-poppy
     # On DUT:
     ectool gpioset PP3300_DX_BASE 0; ectool gpioset PP3300_DX_BASE 1
     # Watch /var/log/messages to ensure hammerd runs correctly

Change-Id: I7f959db68a9bbcb9533dcac35dd6f11894fcaf48
Reviewed-on: https://chromium-review.googlesource.com/689314
Commit-Ready: Joel Kitching <kitching@chromium.org>
Tested-by: Joel Kitching <kitching@chromium.org>
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>