chromeos-config: Define binding for touch settings
There are several types of touch devices use in Chrome OS. We need to
install the correct firmware for these devices and in some cases provide
configuration information (e.g. to the android container). Add a binding
for this with an example and short guide.

This does not currently describe how to allow for variants with
sub-models. This is not currently needed for coral, but we should at least
define how to specify whether a touch device is present or not on a
sub-model.

BUG=b:66994707
TEST=FEATURES=test emerge-coral chromeos-config-tools
Change-Id: Ia60e752c676cfd1be4c290e0b599c3843eb542b9
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/690435