cryptohome: memorize can-attempt-ownership decision
A device can suddenly reboot between the moment Chrome sent
TpmCanAttemptOwnership and the ownership actualy taken. If that
happens, a new TpmCanAttemptOwnership won't be sent at least
until a user logs in (which can be itself prevented by incomplete
tpm initialization). To minimize the window when it can happen
remember the fact that we got TpmCanAttemptOwnership from Chrome
already and re-attempt owning on the next boot, if it was interrupted.

BUG=b:66923649
TEST=interrupt tpm initialization by rebooting, check that it is
     attempted again on the next boot

Change-Id: I3c1ccc4924854a7eeb31f2fd15b21ed4258105dc
Reviewed-on: https://chromium-review.googlesource.com/690581
Commit-Ready: Andrey Pronin <apronin@chromium.org>
Tested-by: Andrey Pronin <apronin@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>