cryptohome: don't reset DA counter on unowned tpm
For TPM2.0 the password for the lockout hierarchy is saved
before the authorization for the hierarchy is actually changed.
As a result, if the initialization is interrupted at the wrong
step, the password may be incorrect, and attempts to reset
DA counter would lead to locking out RH_LOCKOUT.

For TPM1.2, the delegate required for resetting the DA counter
is created on PrepareForEnrollment step.

In both cases it makes sense to wait until the tpm is fully
initialized before attempting ResetDictionaryAttackMitigation().

BUG=b:66923649
TEST=interrupt tpm initialization on tpm2.0 between saving
     the auth values in the local state and actually setting
     the lockout auth value, reboot, verify that continuing
     the interrupted initialization doesn't lead to
     TPM_RC_LOCKOUT.

Change-Id: I132957b64590316fe25d0e74ebfdb0e33507427c
Reviewed-on: https://chromium-review.googlesource.com/691176
Commit-Ready: Andrey Pronin <apronin@chromium.org>
Tested-by: Andrey Pronin <apronin@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>