chromeos-config: Initial JSON support for config
Added initial support for the JSON config format.

Note from Simon:

With this implementation we use either JSON or DTB but not both,
controlled by a 'USE=json' flag. This allows the two to be developed in
parallel without poluting the code with either other's internals. For
example the .dtb implementation includes code to follow phandles which
will not be needed with JSON, since these are expanded out on the host
side.

BUG=b:66950157
TEST=FEATURES=test USE=json emerge-coral chromeos-base/chromeos-config-tools
Still getting a crash with this, needs more work

Change-Id: I26b7fc8d73da7b37d3e265d8a7cb720eb2dacd82
Reviewed-on: https://chromium-review.googlesource.com/732651
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Simon Glass <sjg@chromium.org>
Reviewed-by: C Shapiro <shapiroc@google.com>