intel: ipu3: Use pointer instead of instance in Intel3aClient
The global variable's construct and destruct sequence
in C++ is not fixed, so use pointer could make the construct
and destruct sequence to be fixed.

Add one static function Intel3AClient::release()
which will be called in the PlatformData::deinit() as
the last one.

BUG=b:62358788
TEST=Build and deploy package to poppy and camera works

Change-Id: I4978073c90077e90d61c8172b278d9732959cf88
Signed-off-by: Yang Liang <liang.l.yang@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/714877
Commit-Ready: WeiX Meng <weix.meng@intel.com>
Tested-by: WeiX Meng <weix.meng@intel.com>
Reviewed-by: Hung-yu Wu <hywu@chromium.org>