chromeos-config: Add python cros_config_host and libcros_config_host
Add a cros_config_host.py and libcros_config_host. This is the Python CLI
interface for CrosConfig, as well as the Python API. The objective of the API
is to cleanly insulate the caller from any notion of the backing config. The CLI
is easy, specialized access to that library.

BUG=chromium:771187
TEST=FEATURES=test sudo -E emerge chromeos-base/chromeos-config-tools

Change-Id: I18091f185bed9852e1c416efd310b25af2661663
Reviewed-on: https://chromium-review.googlesource.com/701658
Commit-Ready: Simon Glass <sjg@chromium.org>
Tested-by: Alec Thilenius <athilenius@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>