power: Run clang-format.
Run the following command to reformat the power_manager/
directory:

  clang-format -i -style=file \
    $(find power_manager/ -name '*.h' -o -name '*.cc')

I'd run clang-format against this directory not that long
ago with a slightly different config, so this is mostly just
argument bin-packing changes.

Also add power_manager/ to the list in PRESUBMIT.cfg to
force future changes to be formatted.

BUG=none
TEST=built with FEATURES=test

Change-Id: Ia15c780a277224a821c6d6ed54f162e65eba876f
Reviewed-on: https://chromium-review.googlesource.com/701326
Commit-Ready: Dan Erat <derat@chromium.org>
Tested-by: Dan Erat <derat@chromium.org>
Reviewed-by: Lutz Justen <ljusten@chromium.org>