Add a basic validator for the master configuration
At present there is no explicit validation that the configuration for any
one board matches the binding. This means that errors may not be noticed
until the package which uses the config is tested.

Add a basic validator to get things started. It focussed on the more
complex validation rather that simple checking that things are present.

This version has everything in one CL, including a few small changes to
the Fdt library.

BUG=chromium:761265
TEST=FEATURES=test emerge-coral chromeos-config-tools
PYTHONPATH=~/cosarm ./validate/validate_config \
 ~/cosarm/chroot/build/coral/usr/share/chromeos-config/config.dtb  \
 ~/cosarm/chroot/build/reef-uni/usr/share/chromeos-config/config.dtb \
 README.md
PYTHONPATH=~/cosarm python validate/validate_config_unittest.py

Change-Id: Ib302c8ff2460c2ece498377de1d77c541afc97fd
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/697947
Reviewed-by: Jason Clinton <jclinton@chromium.org>