intel: ipu3: release KBL_AIC before new it in AicLibrary::init
[Issue]
Green circle in the still picture.

[Root Cause]
LSC table from the AIC are all 0,
so the lens shadding correction doesn't work correctly.
Some variables member in KBL_AIC are wrong when
doing still capture, the wrong variables value makes
the LSC table not to be updated.

[Changes]
Release the KBL_AIC instance before generate a new
instance of KBL_AIC when doing init in AicLibrary.
And the behaviour is the same as the local AIC.

BUG=b:62358788
TEST=Build and deploy package to poppy and camera works

Change-Id: Ie888c3a0955a28d94d69d96e1fd0fe20c1b113f8
Signed-off-by: Yang Liang <liang.l.yang@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/714879
Commit-Ready: WeiX Meng <weix.meng@intel.com>
Tested-by: WeiX Meng <weix.meng@intel.com>
Reviewed-by: Hung-yu Wu <hywu@chromium.org>